requirements.txt
python-dotenv
tweepy==3.10.0